deprag/eval/evaluate.py
from typing import List

from tqdm import tqdm

from ..configs.config import DeepRAGConfig
from ..data.loaders import get_dataset
from ..models.agent import DeepRAGAgent
from ..utils.logging import get_logger
from .metrics import compute_exact_match, compute_f1

logger = get_logger(__name__)


def evaluate(config: DeepRAGConfig):
    """Main function to evaluate a trained DeepRAG model."""
    logger.info("Starting Evaluation")

    # Load model, tokenizer, and data
    # agent = DeepRAGAgent(config.model)
    # tokenizer = ...
    test_dataset = get_dataset(config.data)

    predictions = []
    references = []
    retrieval_counts = []

    # for item in tqdm(test_dataset, desc="Evaluating"):
    #     query = item["question"]
    #     reference = item["answer"]

    #     # Generate an answer with the agent
    #     # This would involve the full generation loop with potential retrievals
    #     # response, num_retrievals = agent.generate_with_retrieval(...)

    #     # For now, using placeholders
    #     response = "placeholder prediction"
    #     num_retrievals = 1

    #     predictions.append(response)
    #     references.append(reference)
    #     retrieval_counts.append(num_retrievals)

    # # Compute metrics
    # em_score = compute_exact_match(predictions, references)
    # f1_score = compute_f1(predictions, references)
    # avg_retrievals = sum(retrieval_counts) / len(retrieval_counts)

    # logger.info(f"Evaluation Results:")
    # logger.info(f"  Exact Match: {em_score:.4f}")
    # logger.info(f"  F1 Score: {f1_score:.4f}")
    # logger.info(f"  Avg Retrievals per Query: {avg_retrievals:.2f}")

    logger.warning("Evaluation script is not fully implemented.")
