pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deeprag"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
description = "A framework for End-to-End Differentiable RAG."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "accelerate>=0.21.0",
    "peft>=0.5.0",
    "trl>=0.7.0",
    "hydra-core>=1.3.0",
    "hydra-colorlog>=1.2.0",
    "typer>=0.9.0",
    "rank_bm25",
    "rouge_score",
]

[project.urls]
"Homepage" = "https://github.com/AyhamJo7/deeprag"
"Bug Tracker" = "https://github.com/AyhamJo7/deeprag/issues"

[project.scripts]
deeprag = "cli:app"

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy",
    "pre-commit",
]
viz = [
    "matplotlib",
    "numpy",
]
